Streamlit/requirements.txt
keras==2.9.0
numpy==1.23.1
opencv_python==4.6.0.66
Pillow==9.2.0
streamlit==1.12.2
tensorflow<2.9.1